Revert "fixing gst post"
This reverts commit 0026bb18545ada67169b20cf07383cadae88040c.

+++ Models/Expense.cs
@@ -18,10 +18,10 @@ namespace invoice_system_backend.Models
         public string Date { get; set; }
 
         [Required]
-        public decimal Amount { get; set; }
+        public float Amount { get; set; }
 
         public string Category { get; set; }
 
-        public decimal GstRate { get; set; }
+        public float GstRate { get; set; }
     }
 }
